kloud/rackspace: add tracking of build status
